rename: src/auth/commands.ts > src/auth/utils.ts
We want to create a general utils file related
to auth. The current code falls under 'general
functionality related to auth' so we can simply
rename it.

In future commits we will add more to this.
